refactor: return undefined when error, remove PhoneInput usage since it is impractical to use shared component here.
