--- src/components/phone-input/index.ts
@@ -1,1 +1,0 @@
-export {PhoneInput} from './PhoneInput';
