+++ src/on-behalf-of/types.ts
@@ -1,3 +1,4 @@
 export type GetAccountByPhoneErrorCode =
+  | 'invalid_phone'
   | 'no_associated_account'
   | 'unknown_error';
